added solution to leetcode problem Two Sum
https://leetcode.com/problems/two-sum/
Solution uses a hashmap.
We iterate thorugh the array, using the variable "pair" as the complement for our target. If we find the pair in the hashmap, it means that at some point it was added from the array to the hashmap, therefore we found the 2 numbers that equal our "target",
which would be pair and nums[i], and we return them. Otherwise, if we find no pairs mathing our sum target we return null.