Create #13 Roman to Integer
Created a hashmap with the roman letter and the value it means.
Started iterating the string back to front to calculate the year. If the previous is bigger than the current letter ( V > I ), we add it to the result (res+value). If it is bigger ( I< V ), we subtract it from the result (res-value). After iterating the whole string, we return the final result.